posts/urls.py
from django.urls import path
from rest_framework.routers import DefaultRouter

from .views import FeedApiView, CommentViewSet,MySavedPostView

router = DefaultRouter()

router.register('comments', CommentViewSet, basename='comments')


urlpatterns = [
    path('feed/', FeedApiView.as_view(), name='feed'),
    path('me/saved/', MySavedPostView.as_view(), name='saved-post'),
]
urlpatterns += router.urls