requirements.txt
Django~=5.0
djangorestframework~=3.16.1
celery~=5.6.0
django-prometheus~=2.4.1
prometheus_client~=0.23.1
djangorestframework_simplejwt~=5.5.1